Move UtilsTest to the mil.dds.anet.test.utils package
