+++ src/test/java/mil/dds/anet/test/resources/ReportsResourceTest.java
@@ -69,8 +69,8 @@ import mil.dds.anet.test.client.TaskSearchSortBy;
 import mil.dds.anet.test.client.util.MutationExecutor;
 import mil.dds.anet.test.client.util.QueryExecutor;
 import mil.dds.anet.test.integration.utils.TestApp;
+import mil.dds.anet.test.utils.UtilsTest;
 import mil.dds.anet.utils.DaoUtils;
-import mil.dds.anet.utils.UtilsTest;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
